feat: abort gracefully if libgirepository is missing
